requirements.txt
crewai==0.28.8
crewai_tools==0.1.6
langchain_community==0.0.29
streamlit
google-generativeai
python-dotenv
langchain